[HER-121] Follow-up Rename buttons folder
