+++ src/home/content.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { Title } from 'home/title'
 import { Carousel } from 'home/carousel'
-import { Live } from 'home/buttons'
+import { Live } from 'home/live'
 
 export const Content = () => <>
   <Title />
