+++ src/home/live/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Grid } from '@material-ui/core'
 
-import { NavigationButton } from 'home/buttons/navigation-button'
+import { NavigationButton } from 'home/live/navigation-button'
 
 export const Live = () => <Grid container spacing={'10'} justify={'center'}>
   <NavigationButton path={'/cameras'}>See Cameras</NavigationButton>
